requirements.txt
google-generativeai>=0.8.5
langchain>=0.1.0
langchain-google-genai>=2.0.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
langgraph>=0.0.20
faiss-cpu>=1.7.4
streamlit>=1.32.0
requests>=2.31.0
flask>=3.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
langsmith>=0.1.0
pydantic>=2.6.0